Split avatar migration tests
